add sercurity with models, service classes and login endpoints

+++ backend/src/main/java/click/snekhome/backend/service/NodeService.java
@@ -3,8 +3,13 @@ package click.snekhome.backend.service;
 import click.snekhome.backend.model.Node;
 import click.snekhome.backend.model.NodeData;
 import click.snekhome.backend.repo.NodeRepo;
+import click.snekhome.backend.security.MongoUser;
+import click.snekhome.backend.security.MongoUserDetailService;
+import click.snekhome.backend.security.Role;
+import click.snekhome.backend.security.UserData;
 import click.snekhome.backend.util.ActionType;
 import click.snekhome.backend.util.IdService;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
@@ -12,15 +17,14 @@ import java.util.List;
 
 @Service
 public class NodeService {
-
-    private static final String PLAYERNAME = "playerUnknown";
-
     private final NodeRepo nodeRepo;
     private final IdService idService;
+    private final MongoUserDetailService mongoUserDetailService;
 
-    public NodeService(NodeRepo nodeRepo, IdService idService) {
+    public NodeService(NodeRepo nodeRepo, IdService idService, MongoUserDetailService mongoUserDetailService) {
         this.nodeRepo = nodeRepo;
         this.idService = idService;
+        this.mongoUserDetailService = mongoUserDetailService;
     }
 
     public List<Node> list() {
@@ -28,28 +32,36 @@ public class NodeService {
     }
 
     public Node add(NodeData nodeData) {
-        Node node = new Node(
-                this.idService.generateId(),
-                null,
-                nodeData.name(),
-                0,
-                100,
-                nodeData.coordinates(),
-                Instant.now().getEpochSecond(),
-                0
-        );
-        this.nodeRepo.insert(node);
-        return node;
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        MongoUser user = this.mongoUserDetailService.getUserByUsername(username);
+        if (user.role().equals(Role.ADMIN)) {
+            Node node = new Node(
+                    this.idService.generateId(),
+                    null,
+                    nodeData.name(),
+                    0,
+                    100,
+                    nodeData.coordinates(),
+                    Instant.now().getEpochSecond(),
+                    0
+            );
+            this.nodeRepo.insert(node);
+            return node;
+        } else {
+            throw new RuntimeException("You are not an admin");
+        }
     }
 
     public Node edit(String id, ActionType actionType) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserData userData = this.mongoUserDetailService.getUserData(username);
         Node node = this.nodeRepo.findById(id).orElseThrow();
         Node newNode;
         if (node.ownerId() == null) {
             if (actionType == ActionType.HACK) {
                 newNode = new Node(
                         node.id(),
-                        PLAYERNAME,
+                        userData.id(),
                         node.name(),
                         node.level() + 1,
                         100,
@@ -59,7 +71,7 @@ public class NodeService {
                 );
                 return this.nodeRepo.save(newNode);
             }
-        } else {
+        } else if (node.ownerId().equals(userData.id())){
             if (actionType == ActionType.HACK) {
                 newNode = new Node(
                         node.id(),
@@ -103,6 +115,12 @@ public class NodeService {
     }
 
     public void delete(String id) {
-        this.nodeRepo.deleteById(id);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        MongoUser user = this.mongoUserDetailService.getUserByUsername(username);
+        if (user.role().equals(Role.ADMIN)) {
+            this.nodeRepo.deleteById(id);
+        } else {
+            throw new RuntimeException("You are not an admin");
+        }
     }
 }
